chore: init pkg
