+++ packages/storybook-config/main.d.ts
@@ -1,0 +1,1 @@
+./dist/main.d.ts
