+++ packages/storybook-config/main.js
@@ -1,0 +1,1 @@
+./dist/main.js
