+++ packages/storybook-config/preview.d.ts
@@ -1,0 +1,1 @@
+./dist/preview.d.ts
