+++ packages/storybook-config/preview.js
@@ -1,0 +1,1 @@
+./dist/preview.js
