+++ packages/storybook-config/src/logo.svg.d.ts
