+++ packages/storybook-config/tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": "../../tsconfig.base.client.json",
+	"compilerOptions": {
+		"rootDir": "./src",
+		"outDir": "./dist",
+	},
+	"include": ["./src/"],
+}
