+++ packages/ui-client/package.json
@@ -28,14 +28,6 @@
 		"@rocket.chat/mock-providers": "workspace:^",
 		"@rocket.chat/ui-avatar": "workspace:~",
 		"@rocket.chat/ui-contexts": "workspace:~",
-		"@storybook/addon-a11y": "^8.6.14",
-		"@storybook/addon-actions": "^8.6.14",
-		"@storybook/addon-docs": "^8.6.14",
-		"@storybook/addon-essentials": "^8.6.14",
-		"@storybook/addon-interactions": "^8.6.14",
-		"@storybook/addon-links": "^8.6.14",
-		"@storybook/addon-styling-webpack": "^1.0.1",
-		"@storybook/addon-webpack5-compiler-babel": "^3.0.6",
 		"@storybook/react": "^8.6.14",
 		"@storybook/react-webpack5": "^8.6.14",
 		"@testing-library/react": "~16.0.1",
@@ -54,7 +46,6 @@
 		"react-dom": "~18.3.1",
 		"react-hook-form": "~7.45.4",
 		"storybook": "^8.6.14",
-		"storybook-dark-mode": "^4.0.2",
 		"typescript": "~5.7.2"
 	},
 	"peerDependencies": {
